Perform the game init in *logic*, but it _way_ slows down
The computations get _much_ slower, going from ~2.5secs per update to ~4.7 secs per update! :o

I _guess_ this is because *all* of the `insert` and `delete` logics are
invalidated whenever inserting/deleting something new, and so we must be
recompiling the init logic on every query, even though it's still going
to be false? Yikes!